requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
google-generativeai==0.3.0
Pillow==10.0.0
python-dotenv==1.0.0
